change angle back to values instead of keystrokes

+++ The Game/Net.cpp
@@ -902,7 +902,7 @@ bool Net::getOnline()
 	return online;
 }
 
-void Net::addpMOVEToQueue(int o_id, int angle, float speed, float x, float y, unsigned long framecounter)// this is called by game thread
+void Net::addpMOVEToQueue(int o_id, float angle, float speed, float x, float y, unsigned long framecounter)// this is called by game thread
 {
 	pMOVE packet;
 	packet.id = o_id;
